add basic ARI system. Needs more work as currently just directly links the ailerons and rudders except at high speed and when on ground. Need to implement the longitudinal washout on ARI as well.
Add basic Roll ratio system. This is currently using the same values as the pitch ratio; which is really just a high speed factor - however the aircraft system is more complex than this - but I can't currently find enough details to implement it so this will have to be sufficient until I've got more information.