F-15 : add ground effect coefficients for Cl,Cd,CM : ref NASA-TM-104278, p228
NASA-TM-104278 also states that 600ft/min is the maximum recommeded for the strut; so each gear will fail if the descent rate exceeds 1200ft/min (and the gear is in contact with the ground; so it is possible (and usual) to damage only one side).

There are three new contact points that are used to substitute for damaged gear.