F-15: Stores: mass and position fixes.
Fix the X position of the mass payload - this was incorrectly set as a delta from CG rather than an absolute X position.

This has always been wrong; however with the mass of the MK-84 it became very apparently wrong when trying to land with 6,000 lbs of mass 1.8m from the nose.

Correct mass for missiles.

Added mass for pylons (ADU-552) and rails (LAU 128), and include MK-84 in missile drag. Exact missile drag possibly isn't right as it is based on estimated values.