F-15: Pitch CAS;
- use our own lagged version of Nz;
- separate out proportional gain and integrator (just because that's more like the aircraft, although we could have a PI contoller).
- adjust P based on VC

overall fairly resposive without much in the way of oscillations; can be a little slow to react at certain speeds - but the model has the same
components as the aircraft (proportional value and integrator) and it seems to work.

Integrator needs to be faded out when gear down; haven't figured that out yet.