F-15 revert to using stability axes for aero
The original report AD-A217 366 refers to the aero being in body axes in a few places, but then also confusingly we have the comment "CFX = FORCE IN STABILITY AXES X DIRECTION (CD IN BODY AXIS). Having checked the equations of motion I can't find where the body axes to stability axes transformation is made so my conclusion is that probably the aero is in stability axes. This is further supported by testing that gives us much higher alpha at low speed. For example on approach clean, with 7000lbs of fuel at ISA with 140kts indicated alpha is around 21 degrees and we're looking at the sky. Performance testing for continuous rate turn was unattainable with body axes forces. The VSPAERO model was also in the stability axes and was a fair match to the aero model.

Unless I can find some other data that is more definitive I'm fairly confident that the original aero data is in the stasbility axes - it is simply a confusion in the report.