F-15: fix augmentation illumintion
The illumination due to augmentation is two light point. Due to the scaling of the values these were way too bright. This fixes that and adjust the colour slightly to better match what I think is correct based on photos (usual caveats about photos being unreliable for colour and brightness apply)